moves formatters from master file to individual files
