+++ packages/cms/src/utils/formatters/abs.js
@@ -1,0 +1,8 @@
+/**
+ * Simple Absolute Value
+ */
+function abs(n) {
+  return Math.abs(n);
+}
+
+module.exports = abs;
