+++ packages/cms/src/utils/formatters/commas.js
@@ -1,0 +1,10 @@
+const libs = require("./libs");
+
+/**
+ * Rounds to nearest whole number and adds commas.
+ */
+function commas(n) {
+  return libs.d3.format(",")(Math.round(n));
+}   
+
+module.exports = commas;
