+++ packages/cms/src/utils/formatters/dollar.js
@@ -1,0 +1,9 @@
+/**
+ * Adds a US dollar sign to the beginning of a String or Number.
+ */
+function dollar(n) {
+  if (typeof n === "number") n = formatters.abbreviate(n);       
+  return n.charAt(0) === "-" ? n.replace("-", "-$") : `$${n}`;
+}
+
+module.exports = dollar;
