+++ packages/cms/src/utils/formatters/grewTo.js
@@ -1,0 +1,8 @@
+/**
+ *
+ */
+function grewTo(n) {   
+  return n < 0 ? "declined from" : n > 0 ? "grew to" : "stayed at";
+}
+
+module.exports = grewTo;
