+++ packages/cms/src/utils/formatters/grewWord.js
@@ -1,0 +1,8 @@
+/**
+ * Returns either "grew", "declined", or "stayed" depending on the provided number's sign.
+ */
+function grewWord(n) {  
+  return n < 0 ? "declined" : n > 0 ? "grew" : "stayed";
+}
+
+module.exports = grewWord;
