+++ packages/cms/src/utils/formatters/growing.js
@@ -1,0 +1,8 @@
+/**
+ *
+ */
+function growing(n) {  
+  return n < 0 ? "declining" : "growing"; 
+}
+
+module.exports = growing;
