+++ packages/cms/src/utils/formatters/increasedWord.js
@@ -1,0 +1,8 @@
+/**
+ *
+ */
+function increasedWord(n) {  
+  return n < 0 ? "decreased" : n > 0 ? "increased" : "remained the same";
+}
+
+module.exports = increasedWord;
