+++ packages/cms/src/utils/formatters/increasing.js
@@ -1,0 +1,8 @@
+/**
+ *
+ */
+function increasing(n) {   
+  return n < 0 ? "decreasing" : n > 0 ? "increasing" : "maintaining";
+}
+
+module.exports = increasing;
