+++ packages/cms/src/utils/formatters/longWord.js
@@ -1,0 +1,8 @@
+/**
+ *
+ */
+function longWord(n) {    
+  return n < 0 ? "shorter" : n > 0 ? "longer" : "similar";
+}
+
+module.exports = longWord;
