+++ packages/cms/src/utils/formatters/moreFewerWord.js
@@ -1,0 +1,8 @@
+/**
+ *
+ */
+function moreFewerWord(n) {  
+  return n < 0 ? "fewer" : "more";
+}
+
+module.exports = moreFewerWord;
