+++ packages/cms/src/utils/formatters/moreLess.js
@@ -1,0 +1,8 @@
+/**
+ * Returns either "more" or "less" depending on the provided number's sign.
+ */
+function moreLess(n) {
+  return n < 0 ? "less" : "more";
+}
+
+module.exports = moreLess;
