+++ packages/cms/src/utils/formatters/plural.js
@@ -1,0 +1,8 @@
+/**
+ * Pluralizes a word.
+ */
+function plural(n) {
+  return n.replace(/\w$/g, chr => chr === "y" ? "ies" : `${chr}s`)
+}
+
+module.exports = plural;
