+++ packages/cms/src/utils/formatters/stripHTML.js
@@ -1,0 +1,8 @@
+/**
+ * Removes all HTML tags from a string.
+ */
+function stripHTML(n) {
+  return n.replace(/<[^>]+>/g, "");
+}
+
+module.exports = stripHTML;
